abstract project selector component
